requirements.txt
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
coreschema==0.0.4
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.1
llvmlite==0.40.1
MarkupSafe==2.1.3
packaging==23.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
